LayoutTests/imported/w3c/web-platform-tests/XMLHttpRequest/send-sync-response-event-order-expected.txt


FAIL XMLHttpRequest: The send() method: event order when synchronous flag is set assert_array_equals: property 2, expected "load(12,12,true)" but got "load(0,0,false)"
Recorded events were:1, 4, load(0,0,false), loadend(0,0,false)
Expected events were:1, 4, load(12,12,true), loadend(12,12,true)
